requirement.txt
python-docx
opencv
paddle
docxcompose
PyMuPDF
fitz
